feat(kernels): accept int64 indices in gather kernel
Change the CUDA gather kernel to accept long long (int64) indices
instead of int (int32). This eliminates the need for CPU-side
int64-to-int32 conversion in GPUEngine.Gather, removing the D2H
copy that blocked CUDA graph capture.

+++ internal/cuda/kernels/gather.go
@@ -6,7 +6,7 @@ package kernels
 #cgo LDFLAGS: -L${SRCDIR} -lkernels -lcudart -lstdc++
 #include <cuda_runtime.h>
 
-extern cudaError_t launch_gather(const float* table, const int* indices,
+extern cudaError_t launch_gather(const float* table, const long long* indices,
                                   float* output, int N, int D, int V,
                                   cudaStream_t stream);
 */
@@ -16,12 +16,12 @@ import "unsafe"
 
 // Gather launches the embedding table gather kernel.
 // table: [V, D] embedding table on device.
-// indices: [N] integer indices on device.
+// indices: [N] int64 indices on device.
 // output: [N, D] output on device.
 func Gather(table unsafe.Pointer, indices unsafe.Pointer,
 	output unsafe.Pointer, N, D, V int, s unsafe.Pointer) error { //nolint:gocritic // match CGo API
 	return checkCUDA(C.launch_gather(
-		(*C.float)(table), (*C.int)(indices),
+		(*C.float)(table), (*C.longlong)(indices),
 		(*C.float)(output),
 		C.int(N), C.int(D), C.int(V), stream(s),
 	), "gather")
